Updated Todo
Updated button to match the one in notes. So removed my Fixed Button jsx as it wasn't consistant. Took away the db connection for tonight as I broke it and ran out of time to find out why. So this is the current working state for the TODO list. I will reconnect the DB tommroow and intergrate the adding of the todos and the retrivals.